[skip-ci] Packit: Update downstream task targets
When I added commit b73eec88, I was under the impression that
the packit config for a downstream Fedora branch should only
contain the tasks for that particular branch. But, that's not
quite the case.

For all downstream tasks like `koji builds` and `bodhi updates`,
Packit now uses the config from the `rawhide` branch on dist-git
https://src.fedoraproject.org/rpms/podman/tree/rawhide . So, this means
all downstream tasks including the ones for F38 and F39 which use a
different version of Podman, need to be specified in the packit config
file that lands in the `rawhide` branch.

This commit re-enables koji and bodhi tasks for all Fedora branches.
Enabling F38 and F39 koji and bodhi will still end up building from
the sources in dist-git, so it's not a conflict to have them enabled on
an upstream branch that won't make its way into F38 and F39.

Labelling as `[skip-ci]` as this doesn't need to go through upstream CI.

Thanks to @majamassarini.

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>