CI: fix search-test flakes
Seeing constant e2e test failures today, in search tests.

Reason: tests are searching for "alpine", which is common,
and we're hitting pollution.

Solution: search for "testdigest_v2s<x>", an image in quay
under the libpod namespace. And, in other tests that rely
on docker.io, switch to quay.

Signed-off-by: Ed Santiago <santiago@redhat.com>