[v4.0-rhel] Use a branched runc instead of manual vendor changes
Commit 480a451d7 manually patched runc/libcontainer in vendor, which is
OK short term but creates a maintenance problem down the line.

The manual patching was done in a rush to address CVE-2024-21626.
A fixed runc version (v1.1.12) could not be used as it requires Go 1.17
(since runc 1.1.6) , while this branch here uses Go 1.16.

Instead of manual patching in place, let's use a new runc branch in
projectatomic repo [1], which contains runc 1.1.2 + CVE fix.

Note runc v1.1.2 (rather than v1.1.0 which was used previosly) was
chosen as a base because the CVE backport was already available. As for
what's vendored here, the only difference from v1.1.0 is
https://github.com/opencontainers/runc/commit/8290c4cf58fb.
Meaning, update from v1.1.0 to v1.1.2 should not bring any new issues.

[1] https://github.com/projectatomic/runc/tree/podman-v4.0-rhel

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>