fix remote build isolation on client side
Follow up to 493179be45 which only fixed the issue on the server despite
me trying to fix it on the client as well, with this change here we
always correctly unset the default on the remote client as the
root/rootless status will be wrong.

This means it is enough for users to either have a new server or client
with the bug fix which should make the update process easier.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>