Add a warning if tmpdir/rundir are not on tmpfs
Podman requires several directories to be on tmpfs so we can
properly detect (and clean up after) reboots: the tmpdir (for
Libpod files) and runroot (for c/storage files). We have not done
any validation that they are actual on tmpfs filesystems in the
past, which leads to no end of people wondering why Podman is
broken on their system (where, perhaps, /tmp is not a tmpfs,
which the FHS allows because sanity cannot exist in this world,
and there is no systemd session so /run/user/$UID does not
exist).

This adds an actual warning on launching Podman with such an
unsupported configuration. There's an environment variable if the
user is certain they want to run Podman like this (perhaps having
used systemd-tmpfiles to automatically clean the relevant dir on
reboot), but otherwise we warn every time a command is run with
runroot/tmpdir on a non-tmpfs filesystem.

Signed-off-by: Matt Heon <mheon@redhat.com>