GHA: Fix intermittent workflow error
Periodically, the discussion-lock workflow throws the error: `Resource
not accessible by integration`

This was identified in the
[upstream](https://github.com/dessant/lock-threads)
issue 47, as caused by a version-5 change that adds support for
management of discussions but requires additional permissions
and possibly settings.  Given the low notification traffic from
discussions, old discussions may remain valid for a long while, and are
a useful community-interface:  Disable management of discussions.

Signed-off-by: Chris Evich <cevich@redhat.com>