[v4.0-rhel] CI maintenance mode
This release was originally built around Fedora 35 and golang 1.16.
However the RHEL platform this branch feeds has since moved on to golang
1.17 out of necessity to support security backports.  Given the
difficulty/complexity of updating the CI platform to F36 (golang 1.18),
and the age of this branch, place it into a slimmed-down
"maintenance-only" mode.

Note: This implies a heavier reliance on downstream QE/QA.

Signed-off-by: Chris Evich <cevich@redhat.com>