[v4.0-rhel] CI: Bump CI env. to F36
This branch is in need of a handful of CVE backports which are
incompatible with golang 1.16 shipping in F35.  This is making the
backports very difficult.  Bump the CI VM up to F36 with golang 1.18,
which is representative of what's happening in the RHEL builds.  The
image used was pulled from the CI in the buildah release-1.26 branch.

Also: Disable the upgrade tests since they're non-functional on F36
without significant changes.

Signed-off-by: Chris Evich <cevich@redhat.com>