Revert "Swap out javascript engine"
This reverts commit 1b6e8d73aa666ee3438e97f9d0d8258a71b88e08.

This does not say what error this is supposed to fix. But assuming this
was about "Self-referencing circular pointer" by redoc then see the
following commit "swagger fix infinitive recursion on some types" for a
proper fix.

Our public docs.podman.io page uses redoc so this doesn't fix anything
other than a local run which is not very helpful.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>