[CI:DOCS] Build & upload release artifacts with GitHub Actions
Add a new GitHub Action that builds and uploads release artifacts. This action is triggered by publishing a release on GitHub. The action will only build if the specfic artifact is missing.

This action also triggers the Windows installer action, since the Windows installer action depends on an uploaded artifact.

Note that the action runs on ubuntu-22.04

Signed-off-by: Ashley Cui <acui@redhat.com>