Makefile: remove useless HACK variable in e2e test
This doesn't do anything and I was unable to find anything in the git
history when or why it was added. The HACK var is just adding another
directory to the ginkgo argument list. As the only arg used was the
"./hack" dir which does not contain any go files it does nothing
besides confusing me about its purpose.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>